+++ composeApp/src/commonMain/kotlin/org/company/app/core/common/LayoutType.kt
@@ -1,4 +1,4 @@
-package org.company.app.utils
+package org.company.app.core.common
 
 sealed interface LayoutType {
     object List : LayoutType
