+++ composeApp/src/commonMain/kotlin/org/company/app/presentation/ui/components/channel/community/ChannelCommunity.kt
@@ -18,7 +18,6 @@ import androidx.compose.material.IconButton
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.Comment
-import androidx.compose.material.icons.filled.Comment
 import androidx.compose.material.icons.filled.IosShare
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.filled.ThumbDown
@@ -42,7 +41,7 @@ import org.company.app.domain.model.videos.Youtube
 import org.company.app.theme.LocalThemeIsDark
 import org.company.app.presentation.ui.screens.detail.formatLikes
 import org.company.app.presentation.ui.screens.detail.formatViewComments
-import org.company.app.utils.getFormattedDate
+import org.company.app.core.common.getFormattedDate
 
 @Composable
 fun ChannelCommunity(
