+++ composeApp/src/commonMain/kotlin/org/company/app/presentation/ui/components/common/NoInternet.kt
@@ -24,7 +24,6 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import org.company.app.theme.LocalThemeIsDark
-import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.painterResource
 import youtube_clone.composeapp.generated.resources.Res
