+++ composeApp/src/commonMain/kotlin/org/company/app/presentation/viewmodel/MainViewModel.kt
@@ -14,8 +14,8 @@ import org.company.app.domain.model.comments.Comments
 import org.company.app.domain.model.search.Search
 import org.company.app.domain.model.videos.Youtube
 import org.company.app.domain.usecases.ResultState
-import org.company.app.utils.LayoutInformation
-import org.company.app.utils.UiData
+import org.company.app.core.common.LayoutInformation
+import org.company.app.core.common.UiData
 import sqldelight.db.YoutubeEntity
